refactor(data-values): move date utils
